refactor(register-platforms): move into utils
