+++ src/backend/index.ts
@@ -5,7 +5,7 @@ import { v4 as uuid_v4 } from 'uuid'
 import * as t from 'io-ts'
 import { NextFunction, Request, Response } from 'express'
 import { createAccessToken } from './create-acccess-token'
-import { registerLtiPlatforms } from './register-lti-platforms'
+import { registerLtiPlatforms } from './util/register-lti-platforms'
 import urlJoin from 'url-join'
 import config from '../utils/config'
 import {
